Skip some steps in vagrant bootstrap when possible

+++ script/vagrant_worker_bootstrap.sh
@@ -1,14 +1,16 @@
 #!/usr/bin/env bash
 
-# update the vm
-echo ""
-echo "+-------------------+"
-echo "| Update the System |"
-echo "+-------------------+"
-echo ""
-#apt-get -y upgrade
-# include extra dependecies
-pip install supervisor virtualenv
+# update the vm (first run only)
+if [ ! -f ~/vagrant_worker_bootstrap ]; then
+    echo ""
+    echo "+-------------------+"
+    echo "| Update the System |"
+    echo "+-------------------+"
+    echo ""
+    #apt-get -y upgrade
+    # include extra dependecies
+    pip install supervisor virtualenv
+fi
 
 # setup the supervisor configs
 if [ ! -f /etc/supervisord.conf ]; then
@@ -49,3 +51,6 @@ su - vagrant -c '~/environments/api/bin/pip install -v -r /vagrant/requirements-
 # setup supervisor
 su - vagrant -c 'cp -f /vagrant/script/vagrant_snap_worker.conf ~/supervisor/snap_worker.conf'
 su - vagrant -c 'supervisorctl update'
+
+# touch a file to know that the setup is done
+touch ~/vagrant_worker_bootstrap
